Move struct types into "src/types/struct.ts"
